cli: remove --no-experimental-global-webcrypto flag
