+++ doc/node.1
@@ -192,9 +192,6 @@ Disable experimental support for the WebSocket API.
 .It Fl -no-experimental-global-customevent
 Disable exposition of the CustomEvent on the global scope.
 .
-.It Fl -no-experimental-global-webcrypto
-Disable exposition of the Web Crypto API on the global scope.
-.
 .It Fl -no-experimental-repl-await
 Disable top-level await keyword support in REPL.
 .
