+++ lib/internal/bootstrap/web/exposed-window-or-worker.js
@@ -11,6 +11,7 @@
 const {
   globalThis,
   ObjectDefineProperty,
+  ObjectGetOwnPropertyDescriptor,
 } = primordials;
 
 const {
@@ -19,6 +20,11 @@ const {
   defineReplaceableLazyAttribute,
   exposeLazyInterfaces,
 } = require('internal/util');
+
+const {
+  ERR_INVALID_THIS,
+  ERR_NO_CRYPTO,
+} = require('internal/errors').codes;
 
 // https://html.spec.whatwg.org/multipage/webappapis.html#windoworworkerglobalscope
 const timers = require('timers');
@@ -89,3 +95,29 @@ exposeLazyInterfaces(globalThis, 'internal/deps/undici/undici', ['WebSocket']);
 internalBinding('wasm_web_api').setImplementation((streamState, source) => {
   require('internal/wasm_web_api').wasmStreamingCallback(streamState, source);
 });
+
+// WebCryptoAPI
+if (internalBinding('config').hasOpenSSL) {
+  defineReplaceableLazyAttribute(
+    globalThis,
+    'internal/crypto/webcrypto',
+    ['crypto'],
+    false,
+    function cryptoThisCheck() {
+      if (this !== globalThis && this != null)
+        throw new ERR_INVALID_THIS(
+          'nullish or must be the global object');
+    },
+  );
+  exposeLazyInterfaces(
+    globalThis, 'internal/crypto/webcrypto',
+    ['Crypto', 'CryptoKey', 'SubtleCrypto'],
+  );
+} else {
+  ObjectDefineProperty(globalThis, 'crypto',
+                       { __proto__: null, ...ObjectGetOwnPropertyDescriptor({
+                         get crypto() {
+                           throw new ERR_NO_CRYPTO();
+                         },
+                       }, 'crypto') });
+}
