+++ lib/internal/crypto/webcrypto.js
@@ -5,7 +5,6 @@ const {
   JSONParse,
   JSONStringify,
   ObjectDefineProperties,
-  ObjectDefineProperty,
   ReflectApply,
   ReflectConstruct,
   SafeSet,
@@ -20,10 +19,6 @@ const {
   kWebCryptoCipherEncrypt,
   kWebCryptoCipherDecrypt,
 } = internalBinding('crypto');
-
-const {
-  getOptionValue,
-} = require('internal/options');
 
 const { TextDecoder, TextEncoder } = require('internal/encoding');
 
@@ -1024,18 +1019,6 @@ ObjectDefineProperties(
       value: randomUUID,
     },
   });
-
-if (getOptionValue('--no-experimental-global-webcrypto')) {
-  // For backward compatibility, keep exposing CryptoKey in the Crypto prototype
-  // when using the flag.
-  ObjectDefineProperty(Crypto.prototype, 'CryptoKey', {
-    __proto__: null,
-    enumerable: true,
-    configurable: true,
-    writable: true,
-    value: CryptoKey,
-  });
-}
 
 ObjectDefineProperties(
   SubtleCrypto.prototype, {
