+++ lib/internal/process/pre_execution.js
@@ -12,7 +12,6 @@ const {
   NumberParseInt,
   ObjectDefineProperty,
   ObjectFreeze,
-  ObjectGetOwnPropertyDescriptor,
   SafeMap,
   String,
   StringPrototypeStartsWith,
@@ -36,9 +35,7 @@ const {
 } = require('internal/util');
 
 const {
-  ERR_INVALID_THIS,
   ERR_MANIFEST_ASSERT_INTEGRITY,
-  ERR_NO_CRYPTO,
   ERR_MISSING_OPTION,
   ERR_ACCESS_DENIED,
 } = require('internal/errors').codes;
@@ -107,7 +104,6 @@ function prepareExecution(options) {
   setupNavigator();
   setupWarningHandler();
   setupUndici();
-  setupWebCrypto();
   setupCustomEvent();
   setupCodeCoverage();
   setupDebugEnv();
@@ -339,41 +335,6 @@ function setupNavigator() {
   // https://html.spec.whatwg.org/multipage/system-state.html#the-navigator-object
   exposeLazyInterfaces(globalThis, 'internal/navigator', ['Navigator']);
   defineReplaceableLazyAttribute(globalThis, 'internal/navigator', ['navigator'], false);
-}
-
-// TODO(aduh95): move this to internal/bootstrap/web/* when the CLI flag is
-//               removed.
-function setupWebCrypto() {
-  if (getEmbedderOptions().noBrowserGlobals ||
-      getOptionValue('--no-experimental-global-webcrypto')) {
-    return;
-  }
-
-  if (internalBinding('config').hasOpenSSL) {
-    defineReplaceableLazyAttribute(
-      globalThis,
-      'internal/crypto/webcrypto',
-      ['crypto'],
-      false,
-      function cryptoThisCheck() {
-        if (this !== globalThis && this != null)
-          throw new ERR_INVALID_THIS(
-            'nullish or must be the global object');
-      },
-    );
-    exposeLazyInterfaces(
-      globalThis, 'internal/crypto/webcrypto',
-      ['Crypto', 'CryptoKey', 'SubtleCrypto'],
-    );
-  } else {
-    ObjectDefineProperty(globalThis, 'crypto',
-                         { __proto__: null, ...ObjectGetOwnPropertyDescriptor({
-                           get crypto() {
-                             throw new ERR_NO_CRYPTO();
-                           },
-                         }, 'crypto') });
-
-  }
 }
 
 function setupCodeCoverage() {
